requirements.txt
ddgs
pillow
requests
matplotlib